+++ .gitignore
@@ -15,13 +15,18 @@
 .DS_Store
  
 # Maven
-log/
 target/
 
 # Misc
 bin/
 .cookiecutter_working_folder/
+
+# Ignore these kind of files on all folders, some editors seem to be too eager and create
+# these files/folders
 **/.factorypath
+**/.classpath
+**/.project
+**/target
 
 # Specific to this project
 generated/
