+++ README.md
@@ -43,8 +43,9 @@ There is a lot of boilerplate code associated to building Vaadin portlets for Li
 We first started by using [cookiecutter templates][cookiecutter] to generate a sample Liferay/Vaadin portlet based on [Maven][maven], but we have now created templates for other kinds of Java projects: command-line tools, portal and generic libraries, as well as added detailed documentation for developers.
 
 ## Available templates
-* Vaadin Portlet running on Liferay: this is your common, run-of-the-mill portlet.
+* Vaadin portlet running on Liferay: this is your common, run-of-the-mill portlet.
 * Command-line tools: tools to be used using a console (e.g., [qpostman-cli](https://github.com/qbicsoftware/postman-cli)).
+* Graphical user interface applications: stand-alone JavaFX applications, such as [experiment-graph-gui](https://github.com/qbicsoftware/experiment-graph-gui).
 * Services: these are similar to command-line tools in their structure, but once a service has been started, it stays "active" until shutdown, like our [portlet deployer service](https://github.com/qbicsoftware/portlet-deployer-service).
 * Portlet libraries: libraries using Liferay and Vaadin dependencies (e.g., [portal-utils-lib](https://github.com/qbicsoftware/portal-utils-lib)).
 * Generic Java libraries: libraries that don't have Vaadin or Liferay as dependencies (e.g., [openbis-client-lib](https://github.com/qbicsoftware/openbis-client-lib)).
