+++ cookiecutters/cli/{{ cookiecutter.artifact_id }}/src/test/java/life/qbic/cli/{{ cookiecutter.main_class }}Test.java
@@ -28,9 +28,9 @@ import static org.junit.Assert.assertEquals;
  * Unit tests for {{ cookiecutter.main_class }}Tool.
  */
 public class {{ cookiecutter.main_class }}Test  {
+    // TODO: write unit tests (you do not need to test ToolExecutor, just test the execute() and shutdown() methods of your tool)
 
     @Test
     public void testToolDoesNotOpenPandorasBox() {
-        // TODO: test your implementation as if you actually got paid for it
     }
 }
