+++ cookiecutters/generic-lib/cookiecutter.json
@@ -2,7 +2,7 @@
     "author": "Winnie Pooh",
     "email": "winnie.the.pooh@qbic.uni-tuebingen.de",
     "artifact_id": "sample-lib",
-    "display_name": "Sample Library",
+    "display_name": "Sample generic Library",
     "version": "1.0.0-SNAPSHOT",
     "short_description": "Sample generic library.",
     "copyright_holder": "QBiC"
