--- cookiecutters/generic-lib/{{ cookiecutter.artifact_id }}/src/main/java/life/qbic/utils/text/TextUtilsSample.java
@@ -1,28 +1,0 @@
-package life.qbic.utils.text;
-
-import org.apache.commons.lang3.StringUtils;
-
-/**
- * Sample generic library.
- */
-public class TextUtilsSample  {
-
-    /**
-     * Checks whether the passed argument is a palindrome. Any leading/trailing whitespace will be ignored. 
-     * 
-     * @param word the word to check.
-     * 
-     * @return {@code true} if the passed word is a palindrome, {@code false} otherwise.
-     * 
-     */
-    public boolean isPalindrome(final String word) {
-        boolean isPalindrome = false;
-        if (StringUtils.isBlank(word)) {
-            isPalindrome = true;
-        } else {
-            final String cleanWord = StringUtils.trim(word).toLowerCase();
-            isPalindrome = cleanWord.equals(StringUtils.reverse(cleanWord));
-        }
-        return isPalindrome;
-    }
-}
