+++ cookiecutters/gui/{{ cookiecutter.artifact_id }}/pom.xml
