+++ cookiecutters/gui/{{ cookiecutter.artifact_id }}/src/main/java/life/qbic/cli/{{ cookiecutter.main_class }}Command.java
@@ -16,8 +16,4 @@ public class {{ cookiecutter.main_class }}Command extends AbstractCommand {
     // String url;
     //
     // using package access level for these members will allow you access them within your main and test classes
-    //
-    // IMPORTANT: ToolExecutor requires that this class contains a public constructor that takes no arguments.
-    //            If you need a custom constructor, make sure to provide a no-arguments public constructor.
-    //            See: https://docs.oracle.com/javase/tutorial/java/javaOO/constructors.html
 }
