+++ cookiecutters/gui/{{ cookiecutter.artifact_id }}/src/main/resources/tool.properties
