+++ cookiecutters/gui/{{ cookiecutter.artifact_id }}/src/test/java/life/qbic/cli/{{ cookiecutter.main_class }}Test.java
