+++ cookiecutters/gui/{{ cookiecutter.artifact_id }}/src/test/resources/log4j2.xml
