+++ cookiecutters/gui/{{ cookiecutter.artifact_id }}/src/test/resources/tool.properties
