+++ cookiecutters/portlet/{{ cookiecutter.artifact_id }}/src/test/java/life/qbic/portal/portlet/{{ cookiecutter.main_class }}PortletTest.java
@@ -1,0 +1,23 @@
+package life.qbic.portal.portlet;
+
+import org.junit.Test;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+/**
+ * Tests for {@link {{ cookiecutter.main_class }}Portlet}.
+ */
+public class {{ cookiecutter.main_class }}PortletTest {
+
+    @Test
+    public void mainUIExtendsQBiCPortletUI() {
+        assertTrue("The main UI class must extend life.qbic.portlet.QBiCPortletUI", 
+            QBiCPortletUI.class.isAssignableFrom({{ cookiecutter.main_class }}Portlet.class));
+    }
+
+    @Test
+    public void mainUIIsNotQBiCPortletUI() {
+        assertFalse("The main UI class must be different to life.qbic.portlet.QBiCPortletUI", 
+            QBiCPortletUI.class.equals({{ cookiecutter.main_class }}Portlet.class));
+    }
+}
