+++ cookiecutters/portlet/{{ cookiecutter.artifact_id }}/src/test/resources/portlet.properties
@@ -1,0 +1,5 @@
+include-and-override=${liferay.home}/qbic-ext.properties
+
+# Do not modify the values. These are automatically resolved by maven during the build.
+version=${project.version}
+repository.url=${project.url}
