+++ cookiecutters/service/cookiecutter.json
@@ -1,10 +1,10 @@
 {
     "author": "Winnie Pooh",
     "email": "winnie.the.pooh@qbic.uni-tuebingen.de",
-    "artifact_id": "sample-service",
-    "display_name": "Sample Service",
-    "main_class": "SampleService",
+    "artifact_id": "sample-cli",
+    "display_name": "Sample command-line Tool",
+    "main_class": "Sample",
     "version": "1.0.0-SNAPSHOT",
-    "short_description": "Sample service that replies 'pong' if you say 'ping'.",
+    "short_description": "Command-line utility to...",
     "copyright_holder": "QBiC"
 }
