+++ cookiecutters/service/{{ cookiecutter.artifact_id }}/src/test/resources/tool.properties
@@ -1,3 +1,4 @@
-# Properties set by maven at build time using "resource filtering"
-version=1.2.3-SNAPSHOT
-project.url=http://test.com
+# These properties were set by maven at build time using "resource filtering"
+tool.version=${project.version}
+tool.repo.url=${project.url}
+tool.name=${project.artifactId}
